Split setup code from indexer manager to another service
