+++ packages/node/src/indexer/indexer.module.ts
@@ -14,6 +14,7 @@ import { FetchService } from './fetch.service';
 import { IndexerManager } from './indexer.manager';
 import { MmrService } from './mmr.service';
 import { PoiService } from './poi.service';
+import { ProjectService } from './project.service';
 import { SandboxService } from './sandbox.service';
 import { StoreService } from './store.service';
 
@@ -42,6 +43,7 @@ import { StoreService } from './store.service';
     DynamicDsService,
     PoiService,
     MmrService,
+    ProjectService,
   ],
   exports: [StoreService, MmrService],
 })
